// src/app/model/packages/consumer/hotel-package-detail-res.ts
export class HotelPackageDetailRes {
    id: string;
    hotelId: string;
    roomType: string;
    roomDescription: string;
    customerInformationPack: string;
    adultCount: string;
    childCount: string;
    pricePerNight: number;
    currency: string;
    discount: string;
    breakfast: string;
}